// RydrDriver/src/API.ts
export type ModelOrderConnection = {
  __typename: 'ModelOrderConnection';
  items: Array<Order | null>;
  nextToken?: string | null;
};

export type Order = {
  __typename: 'Order';
  id: string;
  createdAt: string;
  type: string;
  status: string;
  origin: string;
  destination: string;
  originAddress: string;
  destinationAddress: string;
  fare: number;
  scheduleDate: string;
  passengerNumber: string;
  paymentType: string;
  notes: string;
  designate: string;
  userId: string;
  user?: User | null;
  driverId?: string | null;
  driver?: Driver | null;
  updatedAt: string;
};

export type Driver = {
  __typename: 'Driver';
  id: string;
  username: string;
  given_name: string;
  family_name: string;
  phone_number: string;
  currentLat: number;
  currentLng: number;
  heading: number;
  email: string;
  type: string;
  isActive?: boolean | null;
  orders?: ModelOrderConnection | null;
  createdAt: string;
  updatedAt: string;
};
// ...
export type UpdateDriverMutation = {
  updateDriver?: {
    __typename: 'Driver';
    id: string;
    username: string;
    given_name: string;
    family_name: string;
    phone_number: string;
    currentLat: number;
    currentLng: number;
    heading: number;
    email: string;
    type: string;
    isActive?: boolean | null;
    orders?: {
      __typename: 'ModelOrderConnection';
      nextToken?: string | null;
    } | null;
    createdAt: string;
    updatedAt: string;
  } | null;
};
// ...
export type GetDriverQuery = {
  getDriver?: {
    __typename: 'Driver';
    id: string;
    username: string;
    given_name: string;
    family_name: string;
    phone_number: string;
    currentLat: number;
    currentLng: number;
    heading: number;
    email: string;
    type: string;
    isActive?: boolean | null;
    orders?: {
      __typename: 'ModelOrderConnection';
      nextToken?: string | null;
    } | null;
    createdAt: string;
    updatedAt: string;
  } | null;
};